Improved logging for null objects in resource - (Task #394)
- Added an adapter to the resources to listen for changes that may
introduce a null. The adapter will ouput a stack trace to the log and
mention the resource. After that the diagnostics will be updated to
notify the editor.
- Added some code to check the resource on null objects when performing
the updatediagnostics. A diagnostics error will be created in that case.
This diagnostics are used to inform the user by the editors.

---
Task #394: HotFix: Iprove Logging of Model Resource