Editing domain save now locking workspace - (Task #402)
- Save operations from the editing domain are now locking the workspace
to maintain the order of first locking the workspace, than locking
transactions when needed. The assign displine command test case was
running into a deadlock where it was trying to execute a command without
undo, which was most likely interlocking with a builder still holding
the workspace lock. While waiting for a transaction which was locked by
the save of the diting domain.

---
Task #402: Trying to stabilize Code Coverage Reports