Fixing resolution to linked resources - (Task #417)
- Changed the Migration ResourceSet to always hand back
null for eobjects in referenced resources. This way the linked resources
will remain as proxy objects in the currently migrated concept. procies
are not critical for migration.
- Adjusted the ActiveConceptConfigurationElement to correctly link a
concept to existing ones in the repository. Adjustments also consider
uri fragment resolution.
- Adjusted the test cases by fixing a bug in the old test case bailing
out to early. Also added new asserts to check if the proxy objects get
correctly linked to existing concepts.

---
Task #417: Exception when migrating concepts with references