Fixed issue with SWTBot Test case - (Task #671)
- Fixed test case with dialog box being shown on interface removal. The
dialog is shown workspace locked due to the command framework being
used. Therefore waiting for notifications has to be done in a different
place, otherwise there will be an inter-lock between the workspace job
of waiting for notifications and the dialog being executed in the
workspace lock.

---
Task #671: Synchronize Editor Resource Reload with Workspace